added logo to navbar
