+++ frontend/src/components/Navigation/ProfileButton.css
@@ -1,3 +1,16 @@
 li {
     color: white;
 }
+
+.profileMenu {
+    display: inline;
+    font-size: 20px;
+    justify-content: space-between;
+    padding: 5px;
+}
+
+.logout {
+    font-size: 20px;
+    display: inline;
+    background-color: #1c1e1f;
+}
